Improved htc

+++ src/htc/htc_runner.py
@@ -29,57 +29,23 @@ class HTCondorRunner:
         output_dir.mkdir(parents=True, exist_ok=True)
         main_name = "xbox2_main.py"
 
-       # # install requirements
-       # install_requirements = True
-       # if install_requirements:
-       #     env_command = f"cd {work_dir} ;" \
-       #                   #"pip3 install --upgrade pip;" \
-       #                   #"pip3 install --user virtualenv ;" \
-       #                   "virtualenv venv ;" \
-       #                   "source ./venv/bin/activate ;" \
-       #                   "pip3 install -r requirements.txt ;"
-       #     os.system(env_command)
-#
-       # # creating the master bash file
-       # master_bash_filename = htc_dir / "htc_run.sh"
-       # with open(master_bash_filename, 'w') as file:
-       #     try:
-       #         file.write("#!/bin/bash\n")
-       #         file.write(f"cd {work_dir}\n")
-       #         file.write("virtualenv -p /usr/bin/python3 venv \n")
-       #         file.write(f"source {work_dir}/venv/bin/activate\n")
-       #         file.write("ls ./venv/bin/\n")    file.write("python3 -V\n")
-       #         file.write("which python3\n")
-       #         file.write("echo $PYTHONPATH\n")
-       #         file.write(f"python3 {work_dir / main_name} --file_path={work_dir} --output_path={output_dir}")
-       #     except IOError as e:
-       #         print(f"I/O error({e.errno}): {e.strerror}")
-       # os.system(f"chmod +x {master_bash_filename}")
-
-        # install requirements
-        install_requirements = False
-        if install_requirements:  # pip3 install --upgrade pip
-            env_command = f"cd {work_dir} ;" \
-                          "virtualenv venv ;" \
-                          "source ./venv/bin/activate ;" \
-                          "pip3 install -r requirements.txt ;"
-            os.system(env_command)
-
         # creating the master bash file
+        venv_exists = False
+        if "venv" in os.listdir():
+            venv_exists = True
+
         master_bash_filename = htc_dir / "htc_run.sh"
         with open(master_bash_filename, 'w') as file:
             try:
                 file.write("#!/bin/bash\n")
                 file.write(f"cd {work_dir}\n")
-
-                file.write(f"virtualenv venv\n")
-
-                file.write(f"source {work_dir}/venv/bin/activate\n")
-
-                file.write(f"pip3 install -r requirements.txt\n")
-
-                file.write("which python3\n")
-                file.write("echo $PYTHONPATH\n")
+                if venv_exists:
+                    file.write(f"echo \"Virtual environment already exists, delete folder if change necessary\"")
+                    file.write(f"source {work_dir}/venv/bin/activate\n")
+                else:
+                    file.write(f"virtualenv venv\n")
+                    file.write(f"source {work_dir}/venv/bin/activate\n")
+                    file.write(f"pip3 install -r requirements.txt\n")
                 file.write(f"python3 {work_dir / main_name} --file_path={work_dir} --output_path={output_dir}")
             except IOError as e:
                 print(f"I/O error({e.errno}): {e.strerror}")
